Fix tests and update readme with badges

+++ read_seed_test.go
@@ -18,5 +18,5 @@ func TestReadSeed(t *testing.T) {
 	buffer := bytes.NewBuffer([]byte{0, 1, 2, 3, 4, 5, 6, 7, 8, 9})
 	seed, err := ReadSeed(buffer)
 	assert.NoError(t, err)
-	assert.Equal(t, 283686952308238, seed)
+	assert.Equal(t, int64(283686952308238), seed)
 }
